Break Dependency stuff out into it's own class
